requirements.txt
darkflow==1.0.0
fastai==2.5.3
fastbook==0.0.18
Flask==2.0.2
imutils==0.5.4
Keras==2.4.3
matplotlib==3.3.2
numpy==1.19.3
opencv_python==4.5.3.56
Pillow==8.4.0
silence_tensorflow==1.1.1
tensorflow==2.4.0
Werkzeug==2.0.2
